added assessment views
