+++ App/templates/add_semester.html
@@ -1,6 +1,6 @@
 {% extends "admin_layout.html" %}
 
-{% block title %}{% if semester %}Edit Semester{% else %}Add Semester{% endif %}{% endblock %}
+{% block title %}Add Semester{% endblock %}
 
 {% block head %}
 <link rel="stylesheet" href="../static/css/admin1.css">
@@ -56,25 +56,6 @@
     color: white;
   }
   
-  .sem_container {
-    margin-bottom: 1.5rem;
-  }
-  
-  .radio-group {
-    display: flex;
-    gap: 20px;
-    margin-top: 0.5rem;
-  }
-  
-  .radio-option {
-    display: flex;
-    align-items: center;
-  }
-  
-  .radio-option input {
-    margin-right: 5px;
-  }
-  
   .buttons_container {
     display: flex;
     justify-content: center;
@@ -115,53 +96,31 @@
 
 {% block content %}
 <div id="semesterForm">
-  <h2>{% if semester %}Edit Semester{% else %}Add Semester{% endif %}</h2>
+  <h2>Add Semester</h2>
   
-  <form action="{% if semester %}/updateSemester/{{ semester.id }}{% else %}/add_semester{% endif %}" method="POST">
-    {% if semester %}
-    <input type="hidden" name="semester_id" value="{{ semester.id }}">
-    {% endif %}
-    
+  <form action="/add_semester" method="POST">
     <div class="field_containers">
       <label for="start_date" class="input-label">Start Date</label>
-      <input type="date" id="start_date" name="start_date" value="{{ semester.start_date.strftime('%Y-%m-%d') if semester else '' }}" required>
+      <input type="date" id="start_date" name="start_date" required>
     </div>
     
     <div class="field_containers">
       <label for="end_date" class="input-label">End Date</label>
-      <input type="date" id="end_date" name="end_date" value="{{ semester.end_date.strftime('%Y-%m-%d') if semester else '' }}" required>
-    </div>
-    
-    <div class="sem_container">
-      <label for="sem_num" class="input-label">Semester Number</label>
-      <div class="radio-group">
-        <div class="radio-option">
-          <input type="radio" id="sem1" name="sem_num" value="1" {% if semester and semester.sem_num == 1 %}checked{% elif not semester %}checked{% endif %}>
-          <label for="sem1">1</label>
-        </div>
-        <div class="radio-option">
-          <input type="radio" id="sem2" name="sem_num" value="2" {% if semester and semester.sem_num == 2 %}checked{% endif %}>
-          <label for="sem2">2</label>
-        </div>
-        <div class="radio-option">
-          <input type="radio" id="sem3" name="sem_num" value="3" {% if semester and semester.sem_num == 3 %}checked{% endif %}>
-          <label for="sem3">3</label>
-        </div>
-      </div>
+      <input type="date" id="end_date" name="end_date" required>
     </div>
     
     <div class="field_containers">
       <label for="max_assessments" class="input-label">Maximum Weekly Assessments</label>
-      <input type="number" id="max_assessments" name="max_assessments" value="{{ semester.max_assessments if semester else '3' }}" min="1" required>
+      <input type="number" id="max_assessments" name="max_assessments" value="3" min="1" required>
     </div>
     
     <div class="field_containers">
       <label for="constraint_value" class="input-label">Constraint Value (M)</label>
-      <input type="number" id="constraint_value" name="constraint_value" value="{{ semester.constraint_value if semester else '1000' }}" min="1" required>
+      <input type="number" id="constraint_value" name="constraint_value" value="1000" min="1" required>
     </div>
     
     <div class="buttons_container">
-      <button type="submit" class="formBtn">{% if semester %}Update{% else %}Add{% endif %} Semester</button>
+      <button type="submit" class="formBtn">Add Semester</button>
       <a href="/semester" class="formBtn cancelBtn">Cancel</a>
     </div>
   </form>
