+++ App/templates/update_assessment.html
@@ -1,0 +1,296 @@
+{% extends "layout.html" %}
+
+{% block title %}Update Assessment{% endblock %}
+
+{% block head %}
+<link rel="stylesheet" href="../static/css/main.css">
+<style>
+  .form-container {
+    max-width: 800px;
+    margin: 2rem auto;
+    padding: 2rem;
+    background-color: var(--primary-color);
+    border-radius: 8px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+  }
+
+  h2 {
+    text-align: center;
+    margin-bottom: 1.5rem;
+    color: #fff;
+    font-weight: 600;
+    position: relative;
+    padding-bottom: 10px;
+  }
+
+  h2:after {
+    content: '';
+    position: absolute;
+    bottom: 0;
+    left: 50%;
+    transform: translateX(-50%);
+    width: 100px;
+    height: 3px;
+    background-color: var(--tertiary-color);
+    border-radius: 2px;
+  }
+
+  .form-group {
+    margin-bottom: 1.5rem;
+  }
+
+  label {
+    display: block;
+    margin-bottom: 0.5rem;
+    font-weight: 500;
+    color: #ddd;
+  }
+
+  input[type="text"],
+  input[type="number"],
+  select {
+    width: 100%;
+    padding: 10px 15px;
+    border: 1px solid #444;
+    border-radius: 4px;
+    background-color: #2a2e33;
+    color: #fff;
+    font-size: 1rem;
+  }
+
+  input[type="checkbox"] {
+    margin-right: 10px;
+    transform: scale(1.2);
+  }
+
+  .checkbox-label {
+    display: flex;
+    align-items: center;
+    cursor: pointer;
+  }
+
+  .form-row {
+    display: flex;
+    gap: 20px;
+    margin-bottom: 1.5rem;
+  }
+
+  .form-col {
+    flex: 1;
+  }
+
+  .btn-container {
+    display: flex;
+    justify-content: space-between;
+    margin-top: 2rem;
+  }
+
+  .btn {
+    padding: 10px 20px;
+    border: none;
+    border-radius: 4px;
+    font-size: 1rem;
+    cursor: pointer;
+    transition: all 0.2s ease;
+  }
+
+  .btn-primary {
+    background-color: var(--tertiary-color);
+    color: white;
+  }
+
+  .btn-secondary {
+    background-color: #6c757d;
+    color: white;
+  }
+
+  .btn-danger {
+    background-color: #e74c3c;
+    color: white;
+  }
+
+  .btn:hover {
+    opacity: 0.9;
+    transform: translateY(-2px);
+  }
+
+  .error-message {
+    color: #e74c3c;
+    font-size: 0.9rem;
+    margin-top: 0.5rem;
+  }
+
+  .course-info {
+    background-color: #2a2e33;
+    padding: 15px;
+    border-radius: 4px;
+    margin-bottom: 1.5rem;
+    border-left: 4px solid var(--tertiary-color);
+  }
+
+  .course-info p {
+    margin: 5px 0;
+    color: #ddd;
+  }
+
+  .course-info strong {
+    color: #fff;
+  }
+  
+  /* Flash messages */
+  .flash-message {
+    padding: 15px;
+    margin-bottom: 20px;
+    border-radius: 4px;
+    font-weight: 500;
+    animation: fadeIn 0.3s ease-in-out;
+  }
+  
+  .flash-success {
+    background-color: rgba(46, 204, 113, 0.2);
+    border-left: 4px solid #2ecc71;
+    color: #2ecc71;
+  }
+  
+  .flash-error {
+    background-color: rgba(231, 76, 60, 0.2);
+    border-left: 4px solid #e74c3c;
+    color: #e74c3c;
+  }
+  
+  @keyframes fadeIn {
+    from { opacity: 0; transform: translateY(-10px); }
+    to { opacity: 1; transform: translateY(0); }
+  }
+
+  /* Responsive adjustments */
+  @media (max-width: 768px) {
+    .form-row {
+      flex-direction: column;
+      gap: 0;
+    }
+  }
+</style>
+{% endblock %}
+
+{% block content %}
+<div class="form-container">
+  <!-- Display flash messages -->
+  {% with messages = get_flashed_messages(with_categories=true) %}
+    {% if messages %}
+      {% for category, message in messages %}
+        <div class="flash-message flash-{{ category if category != '_' else 'success' }}">
+          {{ message }}
+        </div>
+      {% endfor %}
+    {% endif %}
+  {% endwith %}
+
+  <h2>Update Assessment</h2>
+  
+  <div class="course-info">
+    <p><strong>Course:</strong> {{ assessment.course_code }}</p>
+    {% if semester %}
+    <p><strong>Semester:</strong> {{ semester.name }}</p>
+    {% endif %}
+  </div>
+  
+  <form action="{{ url_for('assessment_views.modify_assessment', id=assessment.id) }}" method="POST">
+    <div class="form-group">
+      <label for="name">Assessment Name</label>
+      <input type="text" id="name" name="name" required value="{{ assessment.name }}">
+    </div>
+    
+    <div class="form-group">
+      <label for="percentage">Percentage (% of total course grade)</label>
+      <input type="number" id="percentage" name="percentage" min="0" max="100" required value="{{ assessment.percentage }}">
+    </div>
+    
+    <div class="form-row">
+      <div class="form-col">
+        <label for="start_week">Start Week</label>
+        <input type="number" id="start_week" name="start_week" min="1" max="15" required value="{{ assessment.start_week }}">
+      </div>
+      
+      <div class="form-col">
+        <label for="start_day">Start Day (1-7, Monday-Sunday)</label>
+        <input type="number" id="start_day" name="start_day" min="1" max="7" required value="{{ assessment.start_day }}">
+      </div>
+    </div>
+    
+    <div class="form-row">
+      <div class="form-col">
+        <label for="end_week">End Week</label>
+        <input type="number" id="end_week" name="end_week" min="1" max="15" required value="{{ assessment.end_week }}">
+      </div>
+      
+      <div class="form-col">
+        <label for="end_day">End Day (1-7, Monday-Sunday)</label>
+        <input type="number" id="end_day" name="end_day" min="1" max="7" required value="{{ assessment.end_day }}">
+      </div>
+    </div>
+    
+    <div class="form-group">
+      <label class="checkbox-label">
+        <input type="checkbox" id="proctored" name="proctored" {% if assessment.proctored %}checked{% endif %}>
+        Proctored Assessment (requires supervision)
+      </label>
+    </div>
+    
+    <div class="btn-container">
+      <a href="/assessments" class="btn btn-secondary">Cancel</a>
+      <button type="submit" class="btn btn-primary">Update Assessment</button>
+    </div>
+  </form>
+  
+  <div style="margin-top: 2rem; border-top: 1px solid #444; padding-top: 1rem;">
+    <form action="{{ url_for('assessment_views.delete_assessment_action', assessment_id=assessment.id) }}" method="POST" id="delete-form">
+      <div class="btn-container">
+        <button type="button" class="btn btn-danger" id="delete-btn">Delete Assessment</button>
+      </div>
+    </form>
+  </div>
+</div>
+
+<script>
+  document.addEventListener('DOMContentLoaded', function() {
+    const form = document.querySelector('form');
+    
+    form.addEventListener('submit', function(e) {
+      const startWeek = parseInt(document.getElementById('start_week').value);
+      const startDay = parseInt(document.getElementById('start_day').value);
+      const endWeek = parseInt(document.getElementById('end_week').value);
+      const endDay = parseInt(document.getElementById('end_day').value);
+      
+      // Basic validation for end date being after start date
+      if (endWeek < startWeek || (endWeek === startWeek && endDay < startDay)) {
+        e.preventDefault();
+        alert('End date must be after start date');
+      }
+    });
+    
+    // Delete confirmation
+    const deleteBtn = document.getElementById('delete-btn');
+    const deleteForm = document.getElementById('delete-form');
+    
+    deleteBtn.addEventListener('click', function() {
+      if (confirm('Are you sure you want to delete this assessment? This action cannot be undone.')) {
+        deleteForm.submit();
+      }
+    });
+    
+    // Auto-hide flash messages after 5 seconds
+    setTimeout(function() {
+      const flashMessages = document.querySelectorAll('.flash-message');
+      flashMessages.forEach(message => {
+        message.style.opacity = '0';
+        message.style.transform = 'translateY(-10px)';
+        message.style.transition = 'opacity 0.5s ease, transform 0.5s ease';
+        setTimeout(() => {
+          message.style.display = 'none';
+        }, 500);
+      });
+    }, 5000);
+  });
+</script>
+{% endblock %}
