+++ App/views/__init__.py
@@ -4,6 +4,8 @@ from .index import index_views
 from .auth import auth_views
 from .admin import admin_views
 from .course import course_views
-views = [staff_views,index_views,auth_views,admin_views,course_views,kris_views]
+from .assessment import assessment_views
+
+views = [staff_views, index_views, auth_views, admin_views, course_views, kris_views, assessment_views]
